requirements.txt
discord.py==2.6.4
python-dotenv==1.0.1
uvicorn==0.29.0
fastapi==0.110.1
pytest==8.3.4
pytest-asyncio==0.24.0
